Refactored all double values to BigDecimal.Closes #14

+++ src/main/java/flagship/domain/utils/tariffs/CanalDueTariff.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Set;
 
@@ -16,11 +17,11 @@ import java.util.Set;
 @NoArgsConstructor
 public class CanalDueTariff extends Tariff {
 
-  private Map<PortArea, Double> canalDuesByPortArea;
-  private Map<ShipType, Double> discountCoefficientByShipType;
-  private Map<PortArea, Double> discountCoefficientsByPortAreaForContainers;
-  private Map<PortArea, Double> discountCoefficientsByPortAreaPerCallCountForContainers;
+  private Map<PortArea, BigDecimal> canalDuesByPortArea;
+  private Map<ShipType, BigDecimal> discountCoefficientByShipType;
+  private Map<PortArea, BigDecimal> discountCoefficientsByPortAreaForContainers;
+  private Map<PortArea, BigDecimal> discountCoefficientsByPortAreaPerCallCountForContainers;
   private Set<ShipType> shipTypesNotEligibleForDiscount;
-  private int callCountThreshold;
-  private Double defaultCallCountDiscountCoefficient;
+  private Integer callCountThreshold;
+  private BigDecimal defaultCallCountDiscountCoefficient;
 }
